PLANET-5523 Hide Covers block if no covers
