+++ assets/src/blocks/Covers/CoversFrontend.js
@@ -1,9 +1,24 @@
 import { Covers } from './Covers';
+import { useCovers } from './useCovers';
 import { getCoversClassName } from './getCoversClassName';
 
 export const CoversFrontend = (attributes) => {
   const { covers_view, cover_type, title, description } = attributes;
   const blockClassName = getCoversClassName(cover_type, covers_view);
+
+  const { covers, loading, loadMoreCovers, row } = useCovers(attributes, document.body.dataset.nro);
+
+  const coversProps = {
+    covers,
+    covers_view,
+    row,
+    loadMoreCovers,
+    cover_type,
+  };
+
+  if (loading || !covers.length) {
+    return null;
+  }
 
   return (
     <section className={blockClassName}>
@@ -13,7 +28,7 @@ export const CoversFrontend = (attributes) => {
       {description &&
         <div class="page-section-description" dangerouslySetInnerHTML={{ __html: description }} />
       }
-      <Covers {...attributes} />
+      <Covers {...coversProps} />
     </section>
   );
 }
